refactor: improve code quality
- Extract CommitItem component to reduce ActivityFeed.tsx size
- Reduce ActivityFeed.tsx from 547 to 409 lines
- Verify code quality with linting and type checking
- Add TECHNICAL_DEBT.md to document remaining issues
- Complete task T029
