+++ TODO.md
@@ -256,7 +256,7 @@ This document outlines the detailed task breakdown for refactoring the GitPulse 
 **T029 - Verify Code Quality**
 - Run linting and type checking
 - Ensure no warnings or errors
-- Status: Not Started
+- Status: Completed [x]
 - Estimate: Small
 - Depends On: T028
